[NEW RBO] Added contant folding (intermediate commit)
