+++ ydb/core/kqp/opt/rbo/kqp_operator.h
@@ -267,6 +267,10 @@ class IOperator {
 
     const TTypeAnnotationNode* GetIUType(TInfoUnit iu);
 
+    virtual TVector<TExpNode::TPtr> GetLambdas() { return {}; }
+
+    virtual void ApplyReplaceMap(TNodeOnNodeOwnedMap map, TRBOContext & ctx) { Y_UNUSED(map); Y_UNUSED_(ctx); }
+
     /***
      * Rename information units of this operator using a specified mapping
      */
@@ -341,6 +345,8 @@ class TOpMap : public IUnaryOperator {
            bool project);
     virtual TVector<TInfoUnit> GetOutputIUs() override;
     virtual TVector<TInfoUnit> GetScalarSubplanIUs(TPlanProps& props) override;
+    virtual TVector<TExpNode::TPtr> GetLambdas() override;
+    virtual void ApplyReplaceMap(TNodeOnNodeOwnedMap map, TRBOContext & ctx) override;
 
     bool HasRenames() const;
     bool HasLambdas() const;
@@ -394,6 +400,8 @@ class TOpFilter : public IUnaryOperator {
     virtual TVector<TInfoUnit> GetOutputIUs() override;
     virtual TVector<TInfoUnit> GetScalarSubplanIUs(TPlanProps& props) override;
     virtual TString ToString(TExprContext& ctx) override;
+    virtual TVector<TExpNode::TPtr> GetLambdas() override;
+    virtual void ApplyReplaceMap(TNodeOnNodeOwnedMap map, TRBOContext & ctx) override;
 
     TVector<TInfoUnit> GetFilterIUs(TPlanProps& props) const;
     TConjunctInfo GetConjunctInfo(TPlanProps& props) const;
@@ -429,6 +437,8 @@ class TOpLimit : public IUnaryOperator {
     virtual TVector<TInfoUnit> GetOutputIUs() override;
     void RenameIUs(const THashMap<TInfoUnit, TInfoUnit, TInfoUnit::THashFunction> &renameMap, TExprContext &ctx) override;
     virtual TString ToString(TExprContext& ctx) override;
+    virtual TVector<TExpNode::TPtr> GetLambdas() override;
+    virtual void ApplyReplaceMap(TNodeOnNodeOwnedMap map, TRBOContext & ctx) override;
 
     TExprNode::TPtr LimitCond;
 };
